+++ CHANGELOG.md
@@ -6,7 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
-- [PATCH] Updated **ButtonIcon** to include `type="button"` ([#1438](https://github.com/optimizely/oui/pull/1438))
+- [Patch] Updated **ButtonIcon** with to force center of icon ([#1439](https://github.com/optimizely/oui/pull/1439))
+- [Patch] Updated **ButtonIcon** to include `type="button"` ([#1438](https://github.com/optimizely/oui/pull/1438))
 
 ## 48.7.3 - 2020-11-09
 - [Patch] Exported **Form** from src/main.js ([#1437](https://github.com/optimizely/oui/pull/1437))